+++ App.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { Platform } from 'react-native';
 import { Provider as PaperProvider } from 'react-native-paper';
 import { enableScreens } from 'react-native-screens';
-import { Provider as StoreProvider } from 'react-redux';
+import useCachedResources from './modules/cached-resources/useCachedResources'
+// import { Provider as StoreProvider } from 'react-redux';
 
 import MainNavigation from './components/MainNavigation';
-import useCachedResources from './modules/cached-resources/useCachedResources';
-import configureStore from './modules/state-management/configure-store';
+// import configureStore from './modules/state-management/configure-store';
 import { theme } from './modules/theme';
 
-const store = configureStore();
+// const store = configureStore();
 
 if (Platform.OS === 'android') {
   enableScreens(true);
@@ -22,10 +22,10 @@ export default function App() {
     return null;
   }
   return (
-    <StoreProvider store={store}>
-      <PaperProvider theme={theme}>
-        <MainNavigation />
-      </PaperProvider>
-    </StoreProvider>
+    // <StoreProvider store={store}>
+    <PaperProvider theme={theme}>
+      <MainNavigation />
+    </PaperProvider>
+    // </StoreProvider>
   );
 }