+++ domains/DataCollection/index.js
@@ -15,7 +15,7 @@ import NewRecordSVG from '../../assets/icons/New-Record-icon.svg';
 import FindResidents from '../../components/FindResidents';
 import Header from '../../components/Header';
 import MapView from '../../components/MapView';
-import { deleteData, getData } from '../../modules/async-storage';
+import { getData } from '../../modules/async-storage';
 import { customFormsQuery } from '../../modules/cached-resources';
 import I18n from '../../modules/i18n';
 import { layout } from '../../modules/theme';
@@ -51,13 +51,12 @@ const DataCollection = ({ navigation }) => {
     getData('currentUser').then((user) => {
       setSurveyingUser(`${user.firstname || ''} ${user.lastname || ''}`);
     });
-
     getData('organization').then((org) => {
       setSurveyingOrganization(org || surveyingOrganization);
     }).catch(() => {
       setSurveyingOrganization(surveyingOrganization || '');
     });
-    customFormsQuery(surveyingOrganization).then((forms) => {
+    getData('customForms').then((forms) => {
       setCustomForms(forms);
     });
   }, [surveyingUser, surveyingOrganization]);
@@ -115,12 +114,13 @@ const DataCollection = ({ navigation }) => {
     });
   };
 
+  const refreshCustomForms = async () => {
+    customFormsQuery(surveyingOrganization).then((forms) => {
+      setCustomForms(forms);
+    });
+  };
   const logOut = () => {
     retrieveSignOutFunction().then(() => {
-      deleteData('credentials');
-      deleteData('pincode');
-      deleteData('organization');
-      deleteData('currentUser');
       navigation.navigate('Sign In');
     });
   };
@@ -224,6 +224,7 @@ const DataCollection = ({ navigation }) => {
                 navigateToCustomForm={navigateToCustomForm}
                 puenteForms={puenteForms}
                 customForms={customForms}
+                refreshCustomForms={refreshCustomForms}
               />
             </View>
           )}