+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [10.5.0](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.4.2...v10.5.0) (2021-02-03)
+### [10.5.4](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.5.3...v10.5.4) (2021-02-04)
+
+### Bug fixes
+
+* inability to refresh custom forms on or offline ([21c85e9](https://github.com/hopetambala/puente-reactnative-collect/commit/21c85e9e20688b0d042bb4dd77b66d8ec4587247))
+* refresh check online status for search bar ([e8fa6a0](https://github.com/hopetambala/puente-reactnative-collect/commit/e8fa6a0e9a282176a13e9af4d8a364fb5d823c04))
+
+### [10.5.1](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.5.0...v10.5.1) (2021-02-04)
+
+### Bug fixes
+
+* customFormsQuery has empty arrays when connected ([23e31cc](https://github.com/hopetambala/puente-reactnative-collect/commit/23e31cc8d393843adab3e8d6040aceeecc65003c))
+* hotfix for pincode offline ([9ff9159](https://github.com/hopetambala/puente-reactnative-collect/commit/9ff915970b5573a26b1d10a6db2bbb13e23abec6))
+* remove deletes ([0b93ddb](https://github.com/hopetambala/puente-reactnative-collect/commit/0b93ddb5e7db8ad688c369253e8fecdd9fe303a9))
+
 
 ### [10.4.2](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.4.1...v10.4.2) (2021-02-01)
 
@@ -11,7 +25,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 * hotfix to have mapview render once ([57066ed](https://github.com/hopetambala/puente-reactnative-collect/commit/57066ed783f723b0c18e784ed565f447ac42eb3f))
 
-### [10.4.1](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.4.0...v10.4.1) (2021-02-01)
 
 ## [10.4.0](https://github.com/hopetambala/puente-reactnative-collect/compare/v10.3.2...v10.4.0) (2021-01-31)